Minor fixes and code cleaning.
* Remove redundant background redrawing RedrawActiveWindow.
* Remove redundant managed cursor callback in RedrawActiveWindow.
* Use refresh to redraw instead of directly calling RedrawActiveWindow.
* Remove unused SetDrawBgColor for drawframe.cpp.
* Fix compiler warning in cvpcb/cvframe.cpp.
* Fix menu spelling and syntax errors in pcbnew.
* Rename Trace_Curseur to DrawCursor in common/drawpanel.cpp.

+++ cvpcb/cvframe.cpp
@@ -200,7 +200,7 @@ WinEDA_CvpcbFrame::WinEDA_CvpcbFrame( const wxString& title,
     if( m_FootprintList )
         m_auimgr.AddPane( m_FootprintList,
                           wxAuiPaneInfo( horiz ).Name( wxT( "m_FootprintList" ) ).
-                          Right().BestSize( m_FrameSize.x * 0.36, m_FrameSize.y ) );
+                          Right().BestSize( (int) ( m_FrameSize.x * 0.36 ), m_FrameSize.y ) );
 
     m_auimgr.Update();
 #endif
